leiningen: remove redundant build steps
Remove build.xml since it builds a broken and unused jar.  Dependencies
for leiningen are not set up properly, so compilation fails, but a jar
is produced anyway because failonerror is not set for the compilation
task.  This doesn't cause failure because the jar is never used - the
lein script downloads the actual jar to the user's home directory on
first run.

Remove all dependencies since they are ignored.  The actual downloaded
leiningen jar is standalone and contains all required dependencies.